Live mode switch now works
Added sequence transaction concept.
Live mode media list marks the first media as a transaction start and the last media as a transaction ends. That way no spacers are added in between sequence transactions.
A delay has been added to the appenderWorkerRunnable when the tryAgain flag is set. This way it leaves time for the rest of the system to load the commandsQueue.
All PccpCommands come with the NO_MODIFIER sequenceModifier by default.